website/backend/requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
SQLAlchemy==2.0.27
requests==2.31.0
python-whois==0.9.3
beautifulsoup4==4.12.2
tldextract==5.1.1
cryptography==41.0.7
ipwhois==1.2.0
python-dotenv==1.0.0
